% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datacite_examples.R
\name{get_datacite_examples}
\alias{get_datacite_examples}
\title{Get example DataCite documents}
\usage{
get_datacite_examples(dir, n)
}
\arguments{
\item{dir}{Local directory to write files to}

\item{n}{Number of example documents from each node to retrieve}
}
\description{
This function grabs a list of example documents for the DataCite dialect from
the KNB, the Arctic Data Center, and Dryad.
}
