% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mn_dialects.R
\name{get_mn_dialects}
\alias{get_mn_dialects}
\title{Get a list of member nodes and dialects}
\usage{
get_mn_dialects()
}
\value{
data.table A summary table of MNs
}
\description{
This function queries the DataONE member nodes and returns a table of
metadata dialects supported by each node.
}
