% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load FAIR data for analysis}
\usage{
load_data()
}
\value{
(data.frame) FAIR data to be analyzed
}
\description{
Load FAIR data for analysis
}
\examples{
fair_data <- load_data()
}
