% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_graphical_theme.R
\name{load_graphical_theme}
\alias{load_graphical_theme}
\title{load_graphical_theme}
\usage{
load_graphical_theme()
}
\value{
theme for ggplot
}
\description{
Load modified ADC theme for ggplot2
}
\examples{
theme_ADC_modified <- load_graphical_theme()
}
